Create separate modules for UndecidableInstances
